Split CanvasInner into 5 hooks (Phase 4b: god-file decomposition)
Problem: web_ui/src/app/canvas/SceneCanvas.tsx was 3324 lines. Most of it
is already well-factored (many exported pure functions at module scope -
toFlowNodes, toFlowEdges, the makeXFns dispatcher factories, geometry
helpers, already independently tested). The actual "god" part was the
CanvasInner component itself, ~993 lines - React Flow's canvas surface,
the single most safety-critical, historically fragile part of this app
(its own comments document a hard-won "DRAG-SYNC REBUILD" fix for a
previously-shipped node/edge desync bug).

Change: extracted 5 hooks into web_ui/src/app/canvas/, matching this
directory's established small-hook convention (useZoomPane.ts,
useLodVisibility.ts, useStreamBuffer.ts):

- useCanvasFontVars.ts - the View-popover font CSS-custom-property effect.
- useBranchFocus.ts - "Hide Other Branches" state + self-healing derivation.
- useViewportReporting.ts - debounced pan/zoom viewport reporting.
- useCanvasPan.ts - the factor-scaled canvas panning subsystem + connection
  hover hit-testing.
- useNodeDragAndSizeSync.ts (452 lines) - THE drag-position-correction
  middleware, node/size-reporting subsystem, onNodesChange, and onDelete.
  Deliberately moved as ONE cohesive unit, not split further: these pieces
  share nodesRef/draggingRef/dragActive and were already tightly
  interdependent by design before this change (the middleware and
  onNodesChange both need the SAME in-flight node mirror to keep drag
  position and edge geometry derived from one number, per the file's own
  DRAG-SYNC REBUILD doc) - artificially separating them for a cosmetic
  line-count win would risk reintroducing exactly the kind of subtle
  desync bug that fix exists to prevent.

CanvasInner itself is now ~483 lines (down from ~993); SceneCanvas.tsx as
a whole is 3324 -> 2823 lines (the 5 new files add 738 lines, mostly
module-doc overhead matching this directory's own convention).

Verification (well beyond the usual bar, given what this component is):
- Every line of the 3 highest-risk relocated functions (the drag-
  correction middleware, onNodesChange, onDelete) was diffed byte-for-byte
  against the pre-refactor original - 100% identical logic; the only
  textual differences anywhere in useNodeDragAndSizeSync.ts vs. the
  original are one added (justified, tested) eslint-disable comment and
  one now-inert eslint-disable comment that turned out to be unnecessary
  in the new file location (confirmed via a direct eslint run: the rule it
  suppressed does not fire either way).
- Full frontend suite (from web_ui/): tsc --noEmit clean, 2033 vitest
  tests passed across 85 files (including SceneCanvas.test.tsx's 195 tests
  covering exactly the drag/smart-guide/size-reporting/panning logic this
  touches), eslint flat at the same pre-existing warning count.
- Started the real dev server (backend + frontend) and drove it against
  the live backend: loaded the sample workspace through the full
  WS pipeline, confirmed real nodes render with real measured DOM
  dimensions (proving the relocated size-reporting code runs correctly
  live, not just in jsdom), and confirmed a real React-Flow-routed
  interaction (the toolbar's Zoom In button) produces the correct
  viewport transform change with zero console errors. Full pixel-level
  drag-and-drop verification via screenshot/coordinate-based interaction
  was not possible in this session - the browser pane could not be
  brought into a composited/displayed state (a known environment
  limitation, not something this specific change caused) - so live drag-
  gesture testing relied on the byte-identical-code guarantee above plus
  the jsdom-level test suite's own drag-middleware coverage, rather than
  a live pixel check. Flagging this gap explicitly rather than claiming a
  visual check that didn't happen.
